pyproject.toml
[project]
name = "infradeployagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.39.4",
    "diagrams>=0.24.4",
    "mcp>=1.11.0",
    "strands-agents>=0.3.0",
    "strands-agents-tools>=0.1.9",
    "streamlit>=1.46.1",
]
